Split query into domain and core layers
