+++ niveshpy/domain/query/__init__.py
@@ -1,0 +1,1 @@
+"""Domain layer for query functionality."""
