+++ niveshpy/infrastructure/sqlite/repositories/sqlite_security_repository.py
@@ -8,7 +8,7 @@ from typing import Any
 from attrs import frozen
 
 from niveshpy.core.logging import logger
-from niveshpy.core.query.ast import Field, FilterNode
+from niveshpy.domain.query.ast import Field, FilterNode
 from niveshpy.exceptions import ResourceNotFoundError
 from niveshpy.infrastructure.sqlite.converters import get_converter
 from niveshpy.infrastructure.sqlite.query import (
