+++ niveshpy/services/account.py
@@ -4,10 +4,10 @@ from collections.abc import Sequence
 from dataclasses import dataclass
 
 from niveshpy.core.logging import logger
-from niveshpy.core.query import ast
 from niveshpy.core.query.prepare import (
     get_prepared_filters_from_queries,
 )
+from niveshpy.domain.query import ast
 from niveshpy.domain.repositories import AccountRepository
 from niveshpy.exceptions import (
     AmbiguousResourceError,
