+++ niveshpy/services/transaction.py
@@ -7,10 +7,10 @@ from dataclasses import dataclass
 
 from attrs import evolve
 
-from niveshpy.core.query import ast
 from niveshpy.core.query.prepare import (
     get_prepared_filters_from_queries,
 )
+from niveshpy.domain.query import ast
 from niveshpy.domain.repositories import (
     AccountRepository,
     SecurityRepository,
