+++ tests/infrastructure/sqlite/repositories/test_sqlite_price_repository.py
@@ -5,7 +5,7 @@ from decimal import Decimal
 
 import pytest
 
-from niveshpy.core.query.ast import Field, FilterNode, Operator
+from niveshpy.domain.query.ast import Field, FilterNode, Operator
 from niveshpy.exceptions import InvalidInputError
 from niveshpy.infrastructure.sqlite.repositories import (
     SqlitePriceRepository,
