+++ tests/infrastructure/sqlite/repositories/test_sqlite_transaction_repository.py
@@ -5,7 +5,7 @@ from decimal import Decimal
 
 import pytest
 
-from niveshpy.core.query.ast import Field, FilterNode, Operator
+from niveshpy.domain.query.ast import Field, FilterNode, Operator
 from niveshpy.infrastructure.sqlite.repositories import (
     SqliteAccountRepository,
     SqliteSecurityRepository,
